renamed functions and files, added transformer functions
